+++ library/core/src/test/java/com/google/android/exoplayer2/upstream/cache/CacheWriterTest.java
@@ -23,14 +23,17 @@ import android.net.Uri;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.android.exoplayer2.C;
+import com.google.android.exoplayer2.testutil.FailOnCloseDataSink;
 import com.google.android.exoplayer2.testutil.FakeDataSet;
 import com.google.android.exoplayer2.testutil.FakeDataSource;
 import com.google.android.exoplayer2.testutil.TestUtil;
 import com.google.android.exoplayer2.upstream.DataSourceException;
 import com.google.android.exoplayer2.upstream.DataSpec;
+import com.google.android.exoplayer2.upstream.FileDataSource;
 import com.google.android.exoplayer2.util.Util;
 import java.io.File;
 import java.io.IOException;
+import java.util.concurrent.atomic.AtomicBoolean;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -213,6 +216,50 @@ public final class CacheWriterTest {
                         /* progressListener= */ null)
                     .cache());
     assertThat(DataSourceException.isCausedByPositionOutOfRange(exception)).isTrue();
+  }
+
+  @Test
+  public void cache_afterFailureOnClose_succeeds() throws Exception {
+    FakeDataSet fakeDataSet = new FakeDataSet().setRandomData("test_data", 100);
+    FakeDataSource upstreamDataSource = new FakeDataSource(fakeDataSet);
+
+    AtomicBoolean failOnClose = new AtomicBoolean(/* initialValue= */ true);
+    FailOnCloseDataSink dataSink = new FailOnCloseDataSink(cache, failOnClose);
+
+    CacheDataSource cacheDataSource =
+        new CacheDataSource(
+            cache,
+            upstreamDataSource,
+            new FileDataSource(),
+            dataSink,
+            /* flags= */ 0,
+            /* eventListener= */ null);
+
+    CachingCounters counters = new CachingCounters();
+
+    CacheWriter cacheWriter =
+        new CacheWriter(
+            cacheDataSource,
+            new DataSpec(Uri.parse("test_data")),
+            /* allowShortContent= */ false,
+            /* temporaryBuffer= */ null,
+            counters);
+
+    // DataSink.close failing must cause the operation to fail rather than succeed.
+    assertThrows(IOException.class, cacheWriter::cache);
+    // Since all of the bytes were read through the DataSource chain successfully before the sink
+    // was closed, the progress listener will have seen all of the bytes being cached, even though
+    // this may not really be the case.
+    counters.assertValues(
+        /* bytesAlreadyCached= */ 0, /* bytesNewlyCached= */ 100, /* contentLength= */ 100);
+
+    failOnClose.set(false);
+
+    // The bytes will be downloaded again, but cached successfully this time.
+    cacheWriter.cache();
+    counters.assertValues(
+        /* bytesAlreadyCached= */ 0, /* bytesNewlyCached= */ 100, /* contentLength= */ 100);
+    assertCachedData(cache, fakeDataSet);
   }
 
   @Test